Initial commit.
This Docker registry is extracted from Reesd. I have been working on this
project on and off since about october last year for Reesd. I think it was
functional quite soon as a proof-of-concept. Then I integrated it with the rest
of the Reesd codebase. Then I started to allow different backends. Indeed you
couldn't do much with my internal Reesd backend. Here there is a simple backend
that uses the file-system.